+++ src/Requests/UpdateRoleRequest.php
@@ -25,6 +25,7 @@ class UpdateRoleRequest extends FormRequest
         return [
             'name'             => ['required', 'string', 'max:255', Rule::unique('roles')->ignore($this->route('id'))],
             'description'      => 'nullable|string|max:255',
+            'type'             => 'nullable|string|max:255',
             'permissions'      => 'nullable|array|min:1',
             'permissions.*'    => 'nullable|integer|exists:permissions,id'
         ];